+++ e2e/index.test.js
@@ -1,0 +1,4 @@
+import { AppRegistry } from 'react-native';
+import { TestableApp } from './src/TestableApp';
+
+AppRegistry.registerComponent('ReactNativeTHEOplayerE2E', () => TestableApp);