+++ e2e/src/tests/index.ts
@@ -1,5 +1,7 @@
-import Basic from './Basic.spec';
 import Ads from './Ads.spec';
+import Basic from './Basic.spec';
 import Connector from './Connector.spec';
+import PresentationMode from './PresentationMode.spec';
+import Version from './Version.spec';
 
-export default [Basic, Ads, Connector];
+export default [Basic, Ads, Connector, PresentationMode, Version];